Can delete runs with alert message. Also add cancel button for new runs.

+++ app/assets/javascripts/controllers/runs/run_controller.js
@@ -4,9 +4,11 @@ BioProjects.RunController = Ember.ObjectController.extend({
       this.get('model').save();
     },
     deleteRun: function(){
-      this.get('model').deleteRecord();
-      this.get('model').save();
-      this.transitionToRoute('runs')
+      if (window.confirm("Are you sure you want to delete this run?")){
+        this.get('model').deleteRecord();
+        this.get('model').save();
+        this.transitionToRoute('runs')
+      }
     }
   }
 });
